services/requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
requests
python-dotenv
psycopg[binary]
gunicorn      
psycopg2-binary